requirements/base.txt
feedparser==6.0.10
lxml==5.1.1
newspaper3k==0.2.8
news-please==1.5.22
beautifulsoup4==4.12.2

langchain==0.0.295
chromadb==0.4.11
openai==0.27.4
tiktoken==0.5.1

# Chroma requires sqlite3 >= 3.35.0 
pysqlite3-binary==0.5.2

# TODO: Need to clean up!
transformers==4.24.0
numpy==1.23.4
torch==1.13.0
sentence-transformers==2.2.2
channels[daphne]==4.0.0
scikit-learn==1.2.2

pandas==1.5.1
matplotlib==3.7.1
plotly==5.14.1

pytz==2022.6  # https://github.com/stub42/pytz
Pillow==9.3.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.2.0  # https://github.com/evansd/whitenoise
redis==4.3.4  # https://github.com/redis/redis-py
hiredis==2.0.0  # https://github.com/redis/hiredis-py
celery==5.2.7  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.4.0  # https://github.com/celery/django-celery-beat
flower==1.2.0  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==4.1.3  # https://www.djangoproject.com/
django-environ==0.9.0  # https://github.com/joke2k/django-environ
django-redis==5.2.0  # https://github.com/jazzband/django-redis
django-import-export==3.2.0
django-public-admin==0.0.5
